Isolate react-app package (#3589)
* Isolate react-app package


---------
