+++ ui/react-app/.gitignore
@@ -1,5 +1,4 @@
 node_modules/
 
-dist/
-
-embed.go
+dist/*
+!dist/*.gz
