+++ ui/react-app/embed.go
@@ -17,3 +17,5 @@ package reactApp
 
 import "embed"
 
+//go:embed dist/26.js.gz dist/26.js.map.gz dist/732.js.gz dist/732.js.map.gz dist/index.html.gz dist/main.js.gz dist/main.js.map.gz
+var embedFS embed.FS
